Rough pass of program's basic functionality
Currently the program communicates with Reddit's API to generate
unique access tokens based on the user's credentials. After a token
is generated, the bearer token is used to authenticate while
communicating with Reddit's API endpoints to pull raw data based
on subreddit keywords determined by the user.

Planned additions will be to properly scale the program allowing
for modularity. Would also like to add features to allow the user
to specify a date range for their search and display the returned
threads as an HTML file. Finally, the user should be able to remove
threads by 'clicking them' after they're done reading them.