Added images and started styling/layout the page
Specifically I focused on getting a style / layout I'm okay with
for the top portion of the web page. This is not a finalized style,
just wanted to get a rough idea first.

I still have to figure out how I want to layout the 2nd half of
the page and how to integrate JS functionality.