Integrated react states to generate unique results
Changed how the program is communicating with the Reddit API
via next JS. Instead of pre-rendering the page with SSR
(Server-side rendering), I transitioned the code over to SSG
(Static-site generation).

This allows the site to remain SEO optimized and hydrate the
page with new threads every time a search is generated.