Added sort filters and began work on new API call
Integrated a drop down menu for users to choose from a list of
sort filters to customize how results are organized.

Removed the search button and replaced the functionality with a key
press associated with the "Enter" key.

Cleaned up overall file layout and made it clear which JS scripts
are specifically for use with Node.

Started working on a new API call that'll pull results from Reddit
based on Subreddits matching the search term rather than all threads
matching the search term.