Cleaned up layout and made site mobile responsive
Cleaned up the layout of how threads are displayed when results are
returned. Specifically added categories for the threads so it's clear
what each portion of the result is corresponding with (i.e. Username,
Thread Name, etc.).

Also integrated media queries to make the web page responsive so any
device can display the content properly.