Added intro page and cleaned component org.
A welcome message was added for when the user first visits the site.
The message includes a brief overview of how to use the program,
how the functionality is designed and planned features for the future.

Also cleaned up the components so the API call function is declared
on a different JS file.