Started incorporating a PDF download feature
Currently the feature is still being tested and I'm unsure if
I'm going to stick with pdf-renderer as it seems to have multiple
issues with next.js. The purpose of this push is to see if it breaks
the main site and if we can properly compile the site off the
main repo branch.