Started working with a new PDF dependency
Able to get a dependency to download pdfs. Now I need to figure
out a way to get the api response to assign it to a variable so we
can save the results. Most likely going to use a different type of
API call to pre-render a response for the pdf function to download.