Finished organizing all existing react components
I also split up the CSS sheets so styling is easier to modify
as we go. The next step is to complete the api call
so that we can finish rebuilding the app.

After the app is ready, the next step is to finish
creating some of the newer features. With the features split up
into individual components, it should be easier to add on additional
features.