Done organizing components and added delete func
Finished refactoring React components and rewrote how API responses
are being mapped to divs. Instead of making the call and mapping
the response within the handleSubmit function, the handleSubmit
function is making an API call and then the returned response
is being mapped to a Thread component which is then wrapping
everything into separate JSX elements.

In addition to organizing everything to be cleaner, this allowed
me to implement a 'delete thread' function. When the api response
is being mapped to the divs, the array indices are being added as
an ID prop which is then used to filter IDs with our delete function.
This allows a delete function to find the ID we're targeting with
an event listener and then updating our 'threads' state with only
the ids that don't match the one we're clicking on. Hence, it deletes
the thread and returns the remaining threads.