Added images to posts
Started tweaking the mobile responsive design. The next step
is to fix some quirks, i.e. the created date doesn't show when the
screen is a normal size. It would be ideal if the page showed the
creation date while the screen was a normal size and was then
resized when the screen is small.

Also need to look into possible means of blurring NSFW images that
show up in results.