Fixed NSFW blur from spilling over div elements
I also ensured that the top nav bar is always at the top of the
page no matter what. This fixed a bug where the blur content would
spill over the nav bar as the user scrolled down.

In addition, I updated the feature list to include remaining
features for the app.